Fix typo

+++ src/deleglise-rivat/pi_deleglise_rivat_parallel1.cpp
@@ -226,7 +226,7 @@ int64_t S2_thread(int64_t x,
 /// Calculate a tiny starting segment_size in order to
 /// evenly balance the work load.
 ///
-int64_t get_segment_size(int128_t x, int64_t limit)
+int64_t get_segment_size(int64_t x, int64_t limit)
 {
   int64_t divisor = ilog(x) * ilog(ilog(x));
   int64_t segment_size = isqrt(limit) / max((int64_t) 1, divisor);
